sorted out old code + working velocity ctrl

+++ app/include/parameter_list.h
@@ -1,6 +1,5 @@
 #pragma once
-#include <string>
-#include <vector>
+
 namespace params {
 int SIM;
 float v;
@@ -11,7 +10,7 @@ float P_vel;
 float I_vel;
 float D_vel;
 float P_pos;
-float T_s;
+int T_s;
 float hover;
 double k_h;
 int T;
